+++ lib/main.dart
@@ -1,3 +1,4 @@
+import 'package:application/qr/view/qr_view.dart';
 import 'package:application/titlebar.dart';
 import 'package:application/userwanderlists/view/userwanderlists.dart';
 import 'package:firebase_auth/firebase_auth.dart';
@@ -7,7 +8,6 @@ import 'package:application/apptheme.dart';
 
 import 'home/view/home_page.dart';
 import 'login/view/login.dart';
-import 'navbar.dart';
 import 'titlebar.dart';
 
 FirebaseAuth auth = FirebaseAuth.instance;
@@ -36,18 +36,29 @@ class MyApp extends StatelessWidget {
       debugShowCheckedModeBanner: false,
     );
   }
+
 }
 
 class MyHomePage extends StatefulWidget {
   MyHomePage({Key? key, required this.title}) : super(key: key);
   final String title;
 
+
   @override
   _MyHomePageState createState() => _MyHomePageState();
 }
 
 class _MyHomePageState extends State<MyHomePage> {
-  @override
+  int _selectedIndex = 0;
+  final PageController _pageController = PageController();
+
+  void _switch(int index) {
+    setState(() {
+      _selectedIndex = index;
+    });
+    _pageController.jumpToPage(index);
+  }
+
   void checkSignedIn() {
     if (true || FirebaseAuth.instance.currentUser == null) {
       Navigator.push(
@@ -55,16 +66,16 @@ class _MyHomePageState extends State<MyHomePage> {
     }
   }
 
+  @override
   Widget build(BuildContext context) {
     return Scaffold(
       backgroundColor: WanTheme.colors.offWhite,
       appBar: Titlebar(),
       body: Center(
-        child: ListView(
-          padding: EdgeInsets.only(
-            left: 20,
-            right: 20,
-          ),
+        child: PageView (
+          // update navbar
+          onPageChanged: (page) => setState(() => _selectedIndex = page),
+          controller: _pageController,
           children: <Widget>[
             Container(child: HomePage()),
             UserWanderlists(),
@@ -105,11 +116,42 @@ class _MyHomePageState extends State<MyHomePage> {
                   },
                 ),
               ),
-            )
+            ),
+            Container(child: WanQrPage()),
           ],
         ),
       ),
-      bottomNavigationBar: Navbar(),
+      bottomNavigationBar: BottomNavigationBar(
+        type: BottomNavigationBarType.fixed, /* No switching animation */
+        showSelectedLabels: false, /* Remove labels */
+        showUnselectedLabels: false,
+        selectedItemColor: WanTheme.colors.pink,
+        unselectedItemColor: WanTheme.colors.grey,
+        items: [
+          BottomNavigationBarItem(
+            label: "Home",
+            icon: Icon(Icons.home_outlined),
+            activeIcon: Icon(Icons.home_rounded),
+          ),
+          BottomNavigationBarItem(
+            label: 'Search',
+            icon: Icon(Icons.search_outlined),
+            activeIcon: Icon(Icons.search_rounded),
+          ),
+          BottomNavigationBarItem(
+            label: 'Flag',
+            icon: Icon(Icons.flag_outlined),
+            activeIcon: Icon(Icons.flag_rounded),
+          ),
+          BottomNavigationBarItem(
+            label: 'Hotel',
+            activeIcon: Icon(Icons.hotel_rounded),
+            icon: Icon(Icons.hotel_outlined),
+          ),
+        ],
+        currentIndex: _selectedIndex,
+        onTap: _switch,
+      ),
     );
   }
 }
